refactor: 쿠키 메서드 분리
CookieUtils 클래스를 생성하여 쿠키 관련 메서드들을 재사용 가능하게 변경
