+++ src/main/java/com/real/backend/oauth/service/AuthService.java
@@ -14,8 +14,9 @@ import com.real.backend.user.domain.LoginType;
 import com.real.backend.user.domain.Role;
 import com.real.backend.user.domain.Status;
 import com.real.backend.user.domain.User;
-import com.real.backend.user.UserRepository;
+import com.real.backend.user.repository.UserRepository;
 import com.real.backend.util.CONSTANT;
+import com.real.backend.util.CookieUtils;
 
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -28,6 +29,7 @@ public class AuthService {
     private final UserRepository userRepository;
     private final KakaoUtil kakaoUtil;
     private final JwtUtil jwtUtil;
+    private final CookieUtils cookieUtils;
 
     public User oAuthLogin(String accessCode, HttpServletResponse response) {
         KakaoTokenDTO oauthToken = kakaoUtil.getAccessToken(accessCode);
@@ -41,21 +43,12 @@ public class AuthService {
         String refreshToken = jwtUtil.generateToken("refresh", user.getId(), user.getNickname(), user.getRole().toString(),
             CONSTANT.REFRESH_TOKEN_EXPIRED);
 
-        ResponseCookie accessCookie = ResponseCookie.from("ACCESS_TOKEN", accessToken)
-            .httpOnly(true)
-            .secure(true)               // HTTPS 필수
-            .path("/")                  // 전 API
-            .sameSite("Lax")            // CSRF 완화
-            .maxAge(CONSTANT.ACCESS_TOKEN_EXPIRED)
-            .build();
 
-        ResponseCookie refreshCookie = ResponseCookie.from("REFRESH_TOKEN", refreshToken)
-            .httpOnly(true)
-            .secure(true)
-            .path("/auth/refresh")      // RT 사용 경로 한정
-            .sameSite("None")           // 크로스-사이트 재발급용
-            .maxAge(CONSTANT.REFRESH_TOKEN_EXPIRED)
-            .build();
+        ResponseCookie accessCookie = cookieUtils.createResponseCookie("ACCESS_TOKEN", accessToken, true, true, "/",
+            "Lax");
+        ResponseCookie refreshCookie = cookieUtils.createResponseCookie("REFRESH_TOKEN", refreshToken, true, true,
+            "/auth/refresh", "None");
+
 
         response.addHeader(HttpHeaders.SET_COOKIE, accessCookie.toString());
         response.addHeader(HttpHeaders.SET_COOKIE, refreshCookie.toString());
